Add support for `fetchPypi`
Add support for updating fetchPypi expressions. This required a bit of
shuffling around because the usual idiom for defining python packages
looks like:

    hdfs = with pythonSuper; (buildPythonPackage rec {
      name = "${pname}-${version}";
      pname = "hdfs";
      version = "2.5.8";

      src = fetchPypi {
        inherit pname version;
        sha256 = "0w4kjjdbcqb7lbgrlbx0vcw3282svwaxrbjipiqmaa61kxa1gq8v";
      };

Specifically, the `inherit` bit there doesn't work with the way this was
previously extracting fields from the expressions - this works now via a
hacky pile of regex searching and narrowing (but hey, it works).